Fix summary console error on charm details page
