+++ templates/details/actions.html
@@ -77,12 +77,6 @@
 </div>
 {% endblock%}
 
-{% block page_scripts %}
-  <script src="{{ versioned_static('js/dist/details.js') }}" defer></script>
+{% block details_scripts %}
   <script src="{{ versioned_static('js/dist/highlight-nav.js') }}" defer></script>
-  <script>
-   window.addEventListener("DOMContentLoaded", function () {
-       charmhub.details.init("{{ package.name }}");
-   });
-  </script>
 {% endblock %}
