+++ templates/details/empty-docs.html
@@ -15,13 +15,3 @@
   </div>
 {% endblock%}
 
-{% block page_scripts %}
-  <script src="{{ versioned_static('js/dist/details.js') }}" defer></script>
-  <script src="{{ versioned_static('js/dist/details_docs.js') }}" defer></script>
-  <script>
-    window.addEventListener("DOMContentLoaded", function () {
-      charmhub.details.init("{{ package.name }}");
-      charmhub.details.docs.initSideNav();
-    });
-  </script>
-{% endblock %}
