+++ templates/details/history.html
@@ -42,7 +42,7 @@
   </div>
 {% endblock%}
 
-{% block page_scripts %}
+{% block details_scripts %}
   <script src="{{ versioned_static('js/dist/details_history.js') }}" defer></script>
   <script>
     window.addEventListener("DOMContentLoaded", function () {
