+++ templates/details/integrate.html
@@ -112,7 +112,7 @@
   </div>
 {% endblock%}
 
-{% block page_scripts %}
+{% block details_scripts %}
   <script src="{{ versioned_static('js/dist/details_integrate.js') }}" defer></script>
   <script>
     /**
